move fvs to subdir detect
