+++ src/ophys_etl/modules/segmentation/detect/feature_vector_rois.py
