+++ tests/modules/segmentation/detect/conftest.py
